refactor: Move function sorting parameters to where it's used
There was no reason to have a method in pkg/project/sort that wraps a toplogysort
call and is only used when resolving parameter values.

Sorting parameters has no relation to sorting projects and should live where it's
actually needed, which is currently the config/parameters/resolve pkg.
