+++ pkg/config/parameter/resolve/sort_test.go
@@ -1,0 +1,184 @@
+/*
+ * @license
+ * Copyright 2023 Dynatrace LLC
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package resolve
+
+import (
+	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/config"
+	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/config/coordinate"
+	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/config/parameter"
+	"github.com/stretchr/testify/assert"
+	"testing"
+)
+
+func TestSortParameters(t *testing.T) {
+	configCoordinates := coordinate.Coordinate{
+		Project:  "project-1",
+		Type:     "dashboard",
+		ConfigId: "dashboard-1",
+	}
+
+	ownerParameterName := "owner"
+	timeoutParameterName := "timeout"
+
+	parameters := config.Parameters{
+		config.NameParameter: &parameter.DummyParameter{
+			References: []parameter.ParameterReference{
+				{
+					Config:   configCoordinates,
+					Property: ownerParameterName,
+				},
+			},
+		},
+		ownerParameterName:   &parameter.DummyParameter{},
+		timeoutParameterName: &parameter.DummyParameter{},
+	}
+
+	sortedParams, errs := sortParameters("", "dev", configCoordinates, parameters)
+
+	assert.Len(t, errs, 0, "expected zero errors when sorting")
+	assert.Equal(t, len(sortedParams), len(parameters), "the same number of parameters should be sorted")
+
+	indexName := indexOfParam(t, sortedParams, config.NameParameter)
+	indexOwner := indexOfParam(t, sortedParams, ownerParameterName)
+
+	assert.Greaterf(t, indexName, indexOwner, "parameter name (index %d) must be after parameter owner (%d)", indexName, indexOwner)
+}
+
+func TestSortParametersShouldFailOnCircularDependency(t *testing.T) {
+	configCoordinates := coordinate.Coordinate{
+		Project:  "project-1",
+		Type:     "dashboard",
+		ConfigId: "dashboard-1",
+	}
+
+	ownerParameterName := "owner"
+
+	parameters := config.Parameters{
+		config.NameParameter: &parameter.DummyParameter{
+			References: []parameter.ParameterReference{
+				{
+					Config:   configCoordinates,
+					Property: ownerParameterName,
+				},
+			},
+		},
+		ownerParameterName: &parameter.DummyParameter{
+			References: []parameter.ParameterReference{
+				{
+					Config:   configCoordinates,
+					Property: config.NameParameter,
+				},
+			},
+		},
+	}
+
+	_, errs := sortParameters("", "dev", configCoordinates, parameters)
+
+	assert.True(t, len(errs) > 0, "should fail")
+}
+
+func indexOfParam(t *testing.T, params []parameter.NamedParameter, name string) int {
+	for i, p := range params {
+		if p.Name == name {
+			return i
+		}
+	}
+
+	t.Fatalf("no parameter with name `%s` found", name)
+	return -1
+}
+
+func TestIsReferencing(t *testing.T) {
+	referencingConfig := coordinate.Coordinate{
+		Project:  "project-1",
+		Type:     "dashboard",
+		ConfigId: "dashboard-1",
+	}
+
+	referencingProperty := "managementZoneName"
+
+	param := parameter.NamedParameter{
+		Name: "name",
+		Parameter: &parameter.DummyParameter{
+			References: []parameter.ParameterReference{
+				{Config: referencingConfig, Property: referencingProperty},
+			},
+		},
+	}
+
+	referencedParameter := parameter.NamedParameter{
+		Name:      referencingProperty,
+		Parameter: &parameter.DummyParameter{},
+	}
+
+	result := parameterReference(param, referencingConfig, referencedParameter)
+
+	assert.True(t, result, "should reference parameter")
+}
+
+func TestIsReferencingShouldReturnFalseForNotReferencing(t *testing.T) {
+	referencingConfig := coordinate.Coordinate{
+		Project:  "project-1",
+		Type:     "dashboard",
+		ConfigId: "dashboard-1",
+	}
+
+	referencingProperty := "managementZoneName"
+
+	param := parameter.NamedParameter{
+		Name: "name",
+		Parameter: &parameter.DummyParameter{
+			References: []parameter.ParameterReference{
+				{
+					Config:   referencingConfig,
+					Property: referencingProperty,
+				},
+			},
+		},
+	}
+
+	referencedParameter := parameter.NamedParameter{
+		Name:      "name",
+		Parameter: &parameter.DummyParameter{},
+	}
+
+	result := parameterReference(param, referencingConfig, referencedParameter)
+
+	assert.True(t, !result, "should not reference parameter")
+}
+
+func TestIsReferencingShouldReturnFalseForParameterWithoutReferences(t *testing.T) {
+	referencingConfig := coordinate.Coordinate{
+		Project:  "project-1",
+		Type:     "dashboard",
+		ConfigId: "dashboard-1",
+	}
+
+	param := parameter.NamedParameter{
+		Name:      "name",
+		Parameter: &parameter.DummyParameter{},
+	}
+
+	referencedParameter := parameter.NamedParameter{
+		Name:      "name",
+		Parameter: &parameter.DummyParameter{},
+	}
+
+	result := parameterReference(param, referencingConfig, referencedParameter)
+
+	assert.True(t, !result, "should not reference parameter")
+}
