+++ pkg/project/v2/sort/sort.go
@@ -19,20 +19,10 @@ package sort
 import (
 	"github.com/dynatrace/dynatrace-configuration-as-code/v2/internal/featureflags"
 	"github.com/dynatrace/dynatrace-configuration-as-code/v2/internal/log"
-	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/config"
-	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/config/coordinate"
-	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/config/parameter"
 	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/graph"
 	project "github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/project/v2"
 	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/project/v2/sort/topologysort"
 )
-
-// Parameters sorts the given parameters of a single configuration in order. If parameters depend on each other, the
-// sorted return slice will contain them in the right order to resolve one after the other.
-// This uses a simple toplogysort implementation.
-func Parameters(group string, environment string, conf coordinate.Coordinate, parameters config.Parameters) ([]parameter.NamedParameter, []error) {
-	return topologysort.SortParameters(group, environment, conf, parameters)
-}
 
 // ConfigsPerEnvironment returns a sorted slice of configurations for each environment. If configurations depend
 // on each other, the slices will contain them in the right order to deploy one after the other.
