+++ pkg/project/v2/sort/sort_test/sort_test.go
@@ -29,84 +29,6 @@ import (
 	"testing"
 )
 
-func TestSortParameters(t *testing.T) {
-	configCoordinates := coordinate.Coordinate{
-		Project:  "project-1",
-		Type:     "dashboard",
-		ConfigId: "dashboard-1",
-	}
-
-	ownerParameterName := "owner"
-	timeoutParameterName := "timeout"
-
-	parameters := config.Parameters{
-		config.NameParameter: &parameter.DummyParameter{
-			References: []parameter.ParameterReference{
-				{
-					Config:   configCoordinates,
-					Property: ownerParameterName,
-				},
-			},
-		},
-		ownerParameterName:   &parameter.DummyParameter{},
-		timeoutParameterName: &parameter.DummyParameter{},
-	}
-
-	sortedParams, errs := sort.Parameters("", "dev", configCoordinates, parameters)
-
-	assert.Len(t, errs, 0, "expected zero errors when sorting")
-	assert.Equal(t, len(sortedParams), len(parameters), "the same number of parameters should be sorted")
-
-	indexName := indexOfParam(t, sortedParams, config.NameParameter)
-	indexOwner := indexOfParam(t, sortedParams, ownerParameterName)
-
-	assert.Greaterf(t, indexName, indexOwner, "parameter name (index %d) must be after parameter owner (%d)", indexName, indexOwner)
-}
-
-func TestSortParametersShouldFailOnCircularDependency(t *testing.T) {
-	configCoordinates := coordinate.Coordinate{
-		Project:  "project-1",
-		Type:     "dashboard",
-		ConfigId: "dashboard-1",
-	}
-
-	ownerParameterName := "owner"
-
-	parameters := config.Parameters{
-		config.NameParameter: &parameter.DummyParameter{
-			References: []parameter.ParameterReference{
-				{
-					Config:   configCoordinates,
-					Property: ownerParameterName,
-				},
-			},
-		},
-		ownerParameterName: &parameter.DummyParameter{
-			References: []parameter.ParameterReference{
-				{
-					Config:   configCoordinates,
-					Property: config.NameParameter,
-				},
-			},
-		},
-	}
-
-	_, errs := sort.Parameters("", "dev", configCoordinates, parameters)
-
-	assert.True(t, len(errs) > 0, "should fail")
-}
-
-func indexOfParam(t *testing.T, params []parameter.NamedParameter, name string) int {
-	for i, p := range params {
-		if p.Name == name {
-			return i
-		}
-	}
-
-	t.Fatalf("no parameter with name `%s` found", name)
-	return -1
-}
-
 func TestGetSortedConfigsForEnvironments(t *testing.T) {
 	projectId := "project1"
 	referencedProjectId := "project2"
